# openapi.yaml
openapi: 3.0.1
info:
  title: Open_Taiwan
  description: check Open government data in Taiwan.
  version: 'v1'
servers:
  - url: http://localhost:3333
paths:
  /reservoirs:
    get:
      operationId: reservoirs
      summary: "Get status Reservoir Resource in Taiwan. Before ask the function, you should ask function:lists to get official full names"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reservoirs'

  /customer_support/creditcard:
    post:
      operationId: cs_cc
      summary: "Get creditcard information in Taiwan By institutions. Before call the function, you must call function:lists and call function:cs_cc_cols to get official full names of columns"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cs_cc'
      requestBody:
        description: the official full names of function:lists and the official full names of columns of function:cs_cc_cols
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                names:
                  required: true
                  type: array
                  items:
                    type: string
                    description: the official full name
                  description: the official full names of function:lists
                columns:
                  required: true
                  type: array
                  items:
                    type: string
                    description: the official full name
                  description: the list of the official full names of function:cs_cc_cols
  
  /customer_support/creditcard/columns:
    post:
      operationId: cs_cc_cols
      summary: "map query subjects to official full names of columns"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cs_cc_cols'

  /lists:
    post:
      operationId: lists
      summary: Get list of official full names with the specific type.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/names'
      requestBody:
        description: Get the list with the specific type
        content:
          application/json:
            schema:
              type: object
              properties:
                type_name:
                  type: string
                  description: the name of type in ['finance','reservoir','other']

parameters:
- name: institutions
  in: path
  description: names of the institutions to use
  required: true
  schema:
    type: array
    items:
      type: string  
  style: simple

components:
  schemas:
    cs_cc:
      type: array
      items: 
        type: object
        properties:
          'no':
            type: string
          'card issuing organization name':
            type: string
          'accounts receivable included in the revolving credit principal, interest commencement date for each account':
            type: string
          'reprinting statement fee':
            type: string
          'check payment dishonored fee':
            type: string
          'clearance (agency) repayment certificate fee':
            type: string
          'foreign transaction settlement fee':
            type: string
          'lost report fee':
            type: string
          "cash advance fee":
            type: string
            description: Cash advance fee
          'review of signed transaction slip fee':
            type: string
          'penalty (late processing fee)':
            type: string
          'overpayment refund fee':
            type: string
          'customer service phone':
            type: string
          'complaint phone':
            type: string
          '24-hour lost report phone':
            type: string
          'head office address':
            type: string
          'website':
            type: string
          'data update year':
            type: string
          'data update month':
            type: string
          'data update day':
            type: string
    reservoirs:
      type: array
      items: 
        type: object
        properties:
          ReservoirName:
            type: string
            description: Name of Reservoir(a.k.a the official full name).
          RecordTime:
            type: string
            description: recoding time.
          WaterLevel:
            type: string
            description: "water level of Reservoir"
        description: List of Reservoirs' status.
    names:
      type: array
      items: 
        type: string
        description: The list of official full names with the specific type
        names:
    cs_cc_cols:
      type: array
      items: 
        type: string
        description: The list of the official full names